:truck: api: move repository
