--- api/lib/infrastructure/serializers/jsonapi/certifications-livret-scolaire/response-objects/Area.js
@@ -1,8 +1,0 @@
-class Area {
-  constructor({ id, name } = {}) {
-    this.id = id;
-    this.name = name;
-  }
-}
-
-export { Area };
