+++ api/src/certification/results/domain/read-models/livret-scolaire/Certificate.js
