+++ api/src/certification/results/domain/read-models/livret-scolaire/CertificateStatus.js
