+++ api/src/certification/results/infrastructure/repositories/competence-tree-repository.js
@@ -1,5 +1,5 @@
-import { CompetenceTree } from '../../../src/shared/domain/models/CompetenceTree.js';
-import * as areaRepository from '../../../src/shared/infrastructure/repositories/area-repository.js';
+import { CompetenceTree } from '../../../../../src/shared//domain/models/CompetenceTree.js';
+import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
 
 const get = async function ({ locale, dependencies = { areaRepository } } = {}) {
   const areas = await dependencies.areaRepository.listWithPixCompetencesOnly({ locale });
