+++ api/src/certification/results/infrastructure/serializers/certification-livret-scolaire-serializer.js
