+++ api/src/certification/results/infrastructure/serializers/factory/build-area-for-livret-scolaire.js
