+++ api/src/certification/results/infrastructure/serializers/factory/build-competences-for-livret-scolaire.js
@@ -1,0 +1,12 @@
+import { Competence } from '../response-objects/Competence.js';
+import { buildArea as buildAreaForLivretScolaire } from './build-area-for-livret-scolaire.js';
+
+const buildCompetenceForLivretScolaire = function ({ id, name, area = buildAreaForLivretScolaire() } = {}) {
+  return new Competence({
+    id,
+    name,
+    area,
+  });
+};
+
+export { buildCompetenceForLivretScolaire };
