+++ api/src/certification/results/infrastructure/serializers/response-objects/Certificate.js
