+++ api/src/certification/results/infrastructure/serializers/response-objects/CertificationsResults.js
