+++ api/src/certification/results/infrastructure/serializers/response-objects/CompetenceResults.js
